Feature/new features

+++ ConventionManager/Controllers/FoodEventController.cs
@@ -230,8 +230,17 @@ namespace ConventionManager.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id, string fromWhere)
         {
             var foodEvent = await _context.FoodEvents.FindAsync(id);
-            _context.FoodEvents.Remove(foodEvent);
-            await _context.SaveChangesAsync();
+            IEnumerable<Subscription> subscriptions = _context.Subscriptions.Where(s => s.Event.Id == id).ToArray();
+            if (subscriptions.Any())
+            {
+                TempData["CannotDeleteEvent"] = foodEvent.CannotDeleteEventMessage;
+            }
+            else
+            {
+                _context.FoodEvents.Remove(foodEvent);
+                await _context.SaveChangesAsync();
+            }
+
             // Checks where the request came from to redirect correctly
             switch (fromWhere)
             {
